requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
streamlit_chat
beautifulsoup4
openai